@@ -15,16 +15,10 @@
   </properties>
 
   <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
       <dependency>
           <groupId>junit</groupId>
           <artifactId>junit</artifactId>
-          <version>4.13</version>
+          <version>4.13.1</version>
           <scope>test</scope>
       </dependency>
     <dependency>
@@ -61,8 +55,8 @@
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-compiler-plugin</artifactId>
             <configuration>
-                <source>9</source>
-                <target>9</target>
+                <source>14</source>
+                <target>14</target>
             </configuration>
         </plugin>
     </plugins>
