@@ -16,20 +16,10 @@ public class BoardFactory {
    * @return The created board object.
    */
   public static Board createBoard(String variant, int marblesPerPlayer, int numOfPlayers) {
-    Board board;
-    switch (variant) {
-      case "standard":
-        board = new StandardBoard(marblesPerPlayer, numOfPlayers);
-        break;
-      case "order":
-        board = new OrderBoard(marblesPerPlayer, numOfPlayers);
-        break;
-      case "yinyang":
-        board = new YinYangBoard(marblesPerPlayer);
-        break;
-      default:
-        board = null;
-    }
-    return board;
+    return switch (variant) {
+      case "order" -> new OrderBoard(marblesPerPlayer, numOfPlayers);
+      case "yinyang" -> new YinYangBoard(marblesPerPlayer);
+      default -> new StandardBoard(marblesPerPlayer, numOfPlayers);
+    };
   }
 }
