@@ -13,12 +13,13 @@ public class CommandHandler {
    * @param message    The message received from the server.
    * @param guiManager The GUI manager to update the user interface.
    */
-  public static void handleCommand(String message, ClientGUIManager guiManager) {
+  public static void handleCommand(String message, ClientGUIManager guiManager) throws InterruptedException {
     if (message.startsWith("User number ") || message.equals("You just moved")) {
       guiManager.clearServerMessageBox();
       guiManager.addCurrPlayerInfo();
     } else if (message.startsWith("Turn skipped by user:") || message.equals("You just skipped")) {
       guiManager.clearServerMessageBox();
+      guiManager.addLabel(message);
       guiManager.addCurrPlayerInfo();
     } else if (message.equals("Game options correct.")) {
       guiManager.setWaitingPanes();
