@@ -12,7 +12,13 @@ public abstract class AbstractBoardManager implements Board {
   protected final int playerZoneHeight;
   protected final int boardHeight; //17
   protected final int boardWidth; //25
-  private int numPlayers;
+  protected final int numOfPlayers;
+  protected int[][] playerZonesStartPoints;
+  protected int numOfCellsPerZone;
+  protected int[] activeZoneNums;
+  protected ArrayList<Integer> finishedPlayers;
+  protected int seed;
+
   protected Cell[][] cells;
 
   protected final int constraintSize;
@@ -23,32 +29,142 @@ public abstract class AbstractBoardManager implements Board {
    * @param marblesPerPlayer The number of marbles per player.
    * @param numOfPlayers     The number of players in the game.
    */
-  public AbstractBoardManager(int marblesPerPlayer, int numOfPlayers) {
+  public AbstractBoardManager(int marblesPerPlayer, int numOfPlayers, int seed) {
     playerZoneHeight = countPlayerZoneHeight(marblesPerPlayer);
     boardHeight = playerZoneHeight * 4 + 1;
     boardWidth = (playerZoneHeight * 6) + 1;
     constraintSize = (1000 / boardWidth) / 2;
-    numPlayers = numOfPlayers;
+    this.numOfPlayers = numOfPlayers;
+    this.seed = seed;
 
     cells = new Cell[boardHeight][boardWidth];
     initializeBoard();
     initializeNeighbors();
 
-    PlayerZoneFactory playerZoneFactory = new PlayerZoneFactory(numOfPlayers, boardWidth, boardHeight, playerZoneHeight);
-    setupPlayerZones(numOfPlayers);
+    this.numOfCellsPerZone = calculateCellsPerZone(playerZoneHeight);
+    this.finishedPlayers = new ArrayList<>();
+    this.playerZonesStartPoints = initializeZoneStartPoints(boardWidth, boardHeight, playerZoneHeight);
+    this.activeZoneNums = new int[6];
+    markActiveZones();
+  }
+
+  protected int[][] initializeZoneStartPoints(int boardWidth, int boardHeight, int playerZoneHeight) {
+    return new int[][]{
+        {0, (boardWidth / 2)},                            // Upper zone
+        {playerZoneHeight * 2 - 1, boardWidth - playerZoneHeight}, // Right upper zone
+        {boardHeight - playerZoneHeight * 2, boardWidth - playerZoneHeight}, // Right bottom zone
+        {boardHeight - 1, (boardWidth / 2)},              // Bottom zone
+        {boardHeight - playerZoneHeight * 2, playerZoneHeight - 1}, // Left bottom zone
+        {playerZoneHeight * 2 - 1, playerZoneHeight - 1}  // Left upper zone
+    };
+  }
+
+  protected int calculateCellsPerZone(int playerZoneHeight) {
+    int counter = playerZoneHeight;
+    int numOfCells = 0;
+    while (counter > 0) {
+      numOfCells += counter;
+      counter--;
+    }
+    return numOfCells;
+  }
+
+  protected void markActiveZones() {
+    Arrays.fill(activeZoneNums, 0);
+    switch (numOfPlayers) {
+      case 2 -> activeZoneNums[0] = activeZoneNums[3] = 1;
+      case 3 -> activeZoneNums[0] = activeZoneNums[2] = activeZoneNums[4] = 1;
+      case 4 -> activeZoneNums[1] = activeZoneNums[2] = activeZoneNums[4] = activeZoneNums[5] = 1;
+      default -> Arrays.fill(activeZoneNums, 1);
+    }
+  }
+
+  protected void setupPlayerZones() {
+    int defaultPlayerNum = 1;
+
+    for (int i = 0; i < playerZonesStartPoints.length; i++) {
+      int[] zoneStartPoint = playerZonesStartPoints[i];
+      int rowStart = zoneStartPoint[0];
+      int colStart = zoneStartPoint[1];
+      int k = 0;
+
+      for (int row = rowStart; checkRow(i, row, rowStart); row = advanceRow(i, row)) {
+        for (int col = colStart - k; col <= colStart + k; col += 2) {
+          setupCell(cells[row][col], i + 1, defaultPlayerNum, activeZoneNums);
+        }
+        k++;
+      }
+
+      if (activeZoneNums[i] == 1) {
+        defaultPlayerNum++;
+      }
+    }
+  }
+
+  protected abstract void setupCell(Cell cell, int zoneNum, int playerNum, int[] activeZoneNums);
+
+  protected void assignCellToZone(Cell cell, int zoneNum, int playerNum) {
+    cell.setZoneNum(zoneNum);
+    cell.setInitialPlayerNum(playerNum);
+  }
+
+  protected void assignPawnToCell(Cell cell, int playerNum) {
+    if (playerNum != 0) {
+      Pawn pawn = new Pawn(playerNum, cell);
+      cell.pawnMoveIn(pawn);
+    }
+  }
+
+  public int checkWinCondition() {
+    for (int i = 0; i < playerZonesStartPoints.length; i++) {
+      int targetPlayer = getTargetPlayer(numOfPlayers, i + 1);
+      if (!finishedPlayers.contains(targetPlayer)) {
+        int[] zoneStartPoint = playerZonesStartPoints[i];
+        int rowStart = zoneStartPoint[0];
+        int colStart = zoneStartPoint[1];
+        int k = 0;
+        int playerPawnCount = 0;
+
+        for (int row = rowStart; checkRow(i, row, rowStart); row = advanceRow(i, row)) {
+          for (int col = colStart - k; col <= colStart + k; col += 2) {
+            if (cells[row][col].getPawn() != null) {
+              if (cells[row][col].getPawn().getPlayerNum() == targetPlayer) {
+                playerPawnCount++;
+              }
+            }
+          }
+          k++;
+        }
+
+        if (playerPawnCount == numOfCellsPerZone) {
+          finishedPlayers.add(targetPlayer);
+          return targetPlayer;
+        }
+      }
+    }
+    return 0;
+  }
+
+  protected abstract int getTargetPlayer(int numOfPlayers, int zoneNum);
+
+  private boolean checkRow(int i, int row, int rowStart) {
+    if (i % 2 == 0) {
+      return row < rowStart + playerZoneHeight;
+    } else {
+      return row > rowStart - playerZoneHeight;
+    }
+  }
+
+  private int advanceRow(int i, int row) {
+    if (i % 2 == 0) {
+      return row + 1;
+    } else {
+      return row - 1;
+    }
   }
 
   /**
-   * Factory method to set up player zones. The specific implementation
-   * is determined by subclasses based on the board type.
-   *
-   * @param numOfPlayers The number of players in the game.
-   */
-  protected abstract void setupPlayerZones(int numOfPlayers);
-
-  /**
-   * @return The size constraint used
-   * for cell rendering in the GUI.
+   * @return The size constraint used for cell rendering in the GUI.
    */
   public int getConstraintSize() {
     return constraintSize;
@@ -171,254 +287,15 @@ public abstract class AbstractBoardManager implements Board {
     return cells[i][j];
   }
 
-  /**
-   * Executes a move by relocating a pawn from a starting cell to a target cell.
-   *
-   * @param input The move details encoded as a string (e.g., "rowStart colStart rowEnd colEnd").
-   * @return A command string describing the move made.
-   */
-  @Override
-  public String makeMove(String input) {
-    int[] moveDetails = decodeInput(input);
-    int rowStart = moveDetails[0];
-    int colStart = moveDetails[1];
-    int rowEnd = moveDetails[3];
-    int colEnd = moveDetails[4];
-
-    Pawn pawn = cells[rowStart][colStart].getPawn();
-    cells[rowStart][colStart].pawnMoveOut();
-    cells[rowEnd][colEnd].pawnMoveIn(pawn);
-    return "[CMD] " + rowStart + " " + colStart + " " + rowEnd + " " + colEnd;
-  }
-
   @Override
   public String makeCreate(int row, int col) {
     if (cells[row][col].getPawn() == null) {
-      return "[CREATE] " + row + " " + col + " " + 0 + " " + numPlayers;
+      return "[CREATE] " + row + " " + col + " " + 0 + " " + numOfPlayers;
     }
     else {
-      return "[CREATE] " + row + " " + col + " " + cells[row][col].getPawn().getPlayerNum() + " " + numPlayers;
+      return "[CREATE] " + row + " " + col + " " + cells[row][col].getPawn().getPlayerNum() + " " + numOfPlayers;
     }
   }
-
-  /**
-   * Validates if a move is legal based on the game rules.
-   *
-   * @param userNum The user number making the move.
-   * @param input   The move details encoded as a string (e.g., "startRow startCol endRow endCol").
-   * @return True if the move is valid; false otherwise.
-   */
-  @Override
-  public boolean validateMove(int userNum, String input) {
-    int[] moveDetails = decodeInput(input);
-    int startRow = moveDetails[0];
-    int startCol = moveDetails[1];
-    int startPlayer = moveDetails[2];
-    int endRow = moveDetails[3];
-    int endCol = moveDetails[4];
-    int endPlayer = moveDetails[5];
-
-    System.out.printf("Player: %d clicked pawn: %d\n", userNum, startPlayer);
-    System.out.printf("Start: [%d, %d], End: [%d, %d]\n", startRow, startCol, endRow, endCol);
-
-    // Validate starting and ending positions
-    if (!isValidStartingPoint(userNum, startPlayer) || !isValidEndingPoint(endPlayer)) {
-      return false;
-    }
-
-    if (cells[startRow][startCol].getPawn() == null) {
-      return false;
-    }
-
-    // Directions for movement
-    int[][] directions = {
-        {-1, -1}, {0, -2}, {1, -1},  // Upper left, left, bottom left
-        {1, 1}, {0, 2}, {-1, 1}      // Bottom right, right, upper right
-    };
-
-    ArrayList<int[]> visited = new ArrayList<>();
-
-    // Check single-step moves
-    if (isValidSingleStep(startRow, startCol, endRow, endCol, directions)) {
-      return true;
-    }
-
-    // Check jumps
-    return isValidJump(startRow, startCol, endRow, endCol, directions, visited);
-  }
-
-  /**
-   * Checks if the starting position of the move belongs to the current player.
-   *
-   * @param userNum    The player number attempting the move.
-   * @param startPlayer The player number associated with the pawn at the start position.
-   * @return True if the starting position is valid; false otherwise.
-   */
-  private boolean isValidStartingPoint(int userNum, int startPlayer) {
-    if (startPlayer != userNum) {
-      System.out.println("Invalid start: Not your pawn.");
-      return false;
-    }
-    return true;
-  }
-
-  /**
-   * Checks if the ending position of the move is unoccupied.
-   *
-   * @param endPlayer The player number at the target position (0 if unoccupied).
-   * @return True if the ending position is valid; false otherwise.
-   */
-  private boolean isValidEndingPoint(int endPlayer) {
-    if (endPlayer != 0) {
-      System.out.println("Invalid end: Target cell is occupied.");
-      return false;
-    }
-    return true;
-  }
-
-  /**
-   * Validates a single-step move by checking if the target cell is adjacent and unoccupied.
-   *
-   * @param startRow   Starting row index.
-   * @param startCol   Starting column index.
-   * @param endRow     Target row index.
-   * @param endCol     Target column index.
-   * @param directions Possible movement directions for the pawn.
-   * @return True if the move is a valid single-step move; false otherwise.
-   */
-  private boolean isValidSingleStep(int startRow, int startCol, int endRow, int endCol, int[][] directions) {
-    for (int[] direction : directions) {
-      int targetRow = startRow + direction[0];
-      int targetCol = startCol + direction[1];
-      if (targetRow == endRow && targetCol == endCol && !cells[targetRow][targetCol].isOccupied()) {
-        System.out.println("Valid single-step move.");
-        return true;
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Validates a jump move by checking if the target cell is reachable via a valid jump.
-   *
-   * @param startRow Starting row index.
-   * @param startCol Starting column index.
-   * @param endRow   Target row index.
-   * @param endCol   Target column index.
-   * @param directions Possible jump directions for the pawn.
-   * @param visited  A list of already visited cells to avoid cycles.
-   * @return True if the move is a valid jump move; false otherwise.
-   */
-  private boolean isValidJump(int startRow, int startCol, int endRow, int endCol, int[][] directions, ArrayList<int[]> visited) {
-    for (int[] direction : directions) {
-      int jumpOverRow = startRow + direction[0];
-      int jumpOverCol = startCol + direction[1];
-
-      // Check if the adjacent cell is within bounds and occupied
-      if (isWithinBounds(jumpOverRow, jumpOverCol) && cells[jumpOverRow][jumpOverCol].isOccupied()) {
-        int landingRow = jumpOverRow + direction[0];
-        int landingCol = jumpOverCol + direction[1];
-
-        // Check if the landing cell is within bounds and unoccupied
-        if (isWithinBounds(landingRow, landingCol) && !cells[landingRow][landingCol].isOccupied()) {
-          if (landingRow == endRow && landingCol == endCol) {
-            System.out.println("Valid jump move.");
-            return true;
-          }
-
-          // Add to visited and recurse
-          visited.add(new int[]{startRow, startCol});
-          if (validateMoveRecursively(landingRow, landingCol, endRow, endCol, visited)) {
-            return true;
-          }
-        }
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Recursively validates multi-jump moves by exploring all possible jump paths.
-   *
-   * @param currentRow Current row index during recursion.
-   * @param currentCol Current column index during recursion.
-   * @param endRow     Target row index.
-   * @param endCol     Target column index.
-   * @param visited    A list of already visited cells to avoid cycles.
-   * @return True if the end cell is reachable via valid jumps; false otherwise.
-   */
-  private boolean validateMoveRecursively(int currentRow, int currentCol, int endRow, int endCol, ArrayList<int[]> visited) {
-    if (cells[currentRow][currentCol].isOccupied()) {
-      System.out.println("Invalid recursive move: Current cell is occupied.");
-      return false;
-    }
-    if (currentRow == endRow && currentCol == endCol) {
-      System.out.println("Recursive move successful.");
-      return true;
-    }
-    if (visited.stream().anyMatch(cell -> Arrays.equals(cell, new int[]{currentRow, currentCol}))) {
-      System.out.println("Invalid recursive move: Already visited this cell.");
-      return false;
-    }
-
-    visited.add(new int[]{currentRow, currentCol});
-    int[][] directions = {
-        {-1, -1}, {0, -2}, {1, -1},  // Upper left, left, bottom left
-        {1, 1}, {0, 2}, {-1, 1}      // Bottom right, right, upper right
-    };
-
-    for (int[] direction : directions) {
-      int jumpOverRow = currentRow + direction[0];
-      int jumpOverCol = currentCol + direction[1];
-      if (isWithinBounds(jumpOverRow, jumpOverCol) && cells[jumpOverRow][jumpOverCol].isOccupied()) {
-        int landingRow = jumpOverRow + direction[0];
-        int landingCol = jumpOverCol + direction[1];
-        if (isWithinBounds(landingRow, landingCol) && !cells[landingRow][landingCol].isOccupied()) {
-          if (validateMoveRecursively(landingRow, landingCol, endRow, endCol, visited)) {
-            return true;
-          }
-        }
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Checks if a cell coordinate is within the bounds of the board.
-   *
-   * @param row The row index.
-   * @param col The column index.
-   * @return True if the cell is within bounds; false otherwise.
-   */
-  private boolean isWithinBounds(int row, int col) {
-    return row >= 0 && row < cells.length && col >= 0 && col < cells[0].length;
-  }
-
-  /**
-   * Decodes a move input string into an array of integers representing the move details.
-   *
-   * @param input The move input as a string.
-   * @return An array of integers containing the parsed move details.
-   */
-  private int[] decodeInput(String input) {
-    String[] tokens = input.split(" ");
-    int[] result = new int[tokens.length];
-    try {
-      for (int i = 0; i < tokens.length; i++) {
-        result[i] = Integer.parseInt(tokens[i]);
-      }
-    } catch (NumberFormatException e) {
-      e.printStackTrace();
-    }
-    return result;
-  }
-
-  /**
-   * Checks the player zones for a win condition.
-   *
-   * @return The player number of the winner, or 0 if no winner is found.
-   */
-  public abstract int checkWin();
 
   /**
    * Retrieves the 2D array of cells representing the board.
@@ -429,7 +306,6 @@ public abstract class AbstractBoardManager implements Board {
   public Cell[][] getCells() {
     return cells;
   }
-
 
 }
 
