@@ -1,11 +1,12 @@
 package org.server.board;
 
+import java.util.Random;
+
 /**
  * Specialized board manager for "Order Out of Chaos" game mode.
  * Configures and validates gameplay for ordered player zones.
  */
 public class OrderBoardManager extends AbstractBoardManager {
-
 
   /**
    * Constructor for the OrderBoardManager class.
@@ -13,29 +14,48 @@ public class OrderBoardManager extends AbstractBoardManager {
    * @param marblesPerPlayer The number of marbles per player.
    * @param numOfPlayers     The number of players in the game.
    */
-  public OrderBoardManager(int marblesPerPlayer, int numOfPlayers) {
-    super(marblesPerPlayer, numOfPlayers);
-
+  public OrderBoardManager(int marblesPerPlayer, int numOfPlayers, int seed) {
+    super(marblesPerPlayer, numOfPlayers, seed);
+    setupPlayerZones();
+    distributePawns();
   }
 
-  /**
-   * Configures player zones specific to the "Order Out of Chaos" mode.
-   *
-   * @param numOfPlayers The number of players in the game.
-   */
   @Override
-  protected void setupPlayerZones(int numOfPlayers) {
-    PlayerZoneFactory playerZoneFactory = new PlayerZoneFactory(numOfPlayers,boardWidth,boardHeight,playerZoneHeight);
-    cells = playerZoneFactory.addOrderPlayerZones(cells);
+  protected void setupCell(Cell cell, int zoneNum, int defaultPlayerNum, int[] activeZoneNums) {
+    int playerNum;
+    if (activeZoneNums[zoneNum - 1] == 1) {
+      playerNum = defaultPlayerNum;
+    } else {
+      playerNum = 0;
+    }
+    assignCellToZone(cell, zoneNum, playerNum);
   }
 
-  /**
-   * Checks for a winning condition in the "Order Out of Chaos" mode.
-   *
-   * @return The player number of the winner, or 0 if no winner is determined.
-   */
+  private void distributePawns() {
+    // Randomly distribute pawns in the middle of the board.
+    System.out.println(seed);
+    Random random = new Random(seed);
+    int player = 0;
+    for (int j : activeZoneNums) {
+      if (j!=0) {
+        player++;
+        int i = 0;
+        while (i < numOfCellsPerZone) {
+          int row = random.nextInt(boardHeight);
+          int col = random.nextInt(boardWidth);
+
+          if (cells[row][col].getPawn() == null && cells[row][col].getZoneNum()==0 && cells[row][col].isInsideBoard()) {
+            assignPawnToCell(cells[row][col], player);
+            System.out.println("row: " + row + ", col: " + col + ", isOccupied: " + cells[row][col].isOccupied());
+            i++;
+          }
+        }
+      }
+    }
+  }
+
   @Override
-  public int checkWin() {
-    return PlayerZoneFactory.checkZoneForWin(cells,"order");
+  protected int getTargetPlayer(int numOfPlayers, int zoneNum) {
+    return TargetPlayerHandler.getTargetPlayerNum(numOfPlayers, zoneNum);
   }
 }
